Extracted hex keyboard implementation to separate module
