+++ app/src/main/res/layout-port/activity_edit_deserializer.xml
@@ -139,7 +139,7 @@
         </androidx.constraintlayout.widget.ConstraintLayout>
     </androidx.core.widget.NestedScrollView>
 
-    <view class="com.aconno.blesniffer.ui.HexKeyboardView"
+    <view class="com.aconno.hexinputlib.HexKeyboardView"
         android:id="@+id/hex_keyboard"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
