+++ gradle.properties
@@ -18,3 +18,7 @@ org.gradle.jvmargs=-Xmx1536m
 # http://www.gradle.org/docs/current/userguide/multi_project_builds.html#sec:decoupled_projects
 # org.gradle.parallel=true
 sensorics_fabricKey=96f02db829648015129ff625e4c95e3408e05e70
+ACONNO_RELEASE_KEYSTORE_PATH=/
+ACONNO_RELEASE_KEYSTORE_PASSWORD=
+SENSORICS_RELEASE_KEY_ALIAS=
+SENSORICS_RELEASE_KEY_PASSWORD=
